+++ trunk/src/main/java/com/compomics/util/experiment/biology/AminoAcidSequence.java
@@ -317,7 +317,7 @@ public class AminoAcidSequence extends ExperimentObject implements TagComponent 
         for (Double mass : masses1.keySet()) {
             Integer occurrence1 = masses1.get(mass);
             Integer occurrence2 = masses2.get(mass);
-            if (occurrence2 == null || occurrence2 != occurrence1) {
+            if (occurrence2 == null || occurrence2.intValue() != occurrence1) {
                 return false;
             }
         }
@@ -837,7 +837,7 @@ public class AminoAcidSequence extends ExperimentObject implements TagComponent 
                 boolean found = false;
                 for (ModificationMatch modificationMatch2 : mods2) {
                     PTM ptm2 = ptmFactory.getPTM(modificationMatch2.getTheoreticPtm());
-                    if (ptm1.getMass() == ptm2.getMass()) {
+                    if (ptm1.getMass() == ptm2.getMass()) { // @TODO: compare against the accuracy
                         found = true;
                         break;
                     }
@@ -903,7 +903,7 @@ public class AminoAcidSequence extends ExperimentObject implements TagComponent 
         for (Double mass : masses1.keySet()) {
             Integer occurrence1 = masses1.get(mass);
             Integer occurrence2 = masses2.get(mass);
-            if (occurrence2 == null || occurrence2 != occurrence1) {
+            if (occurrence2 == null || occurrence2.intValue() != occurrence1) {
                 return false;
             }
         }
